refactor(extension-common): 広告リンクのクリック検知を link-verification へ移す
content-script.ts が、フレームの足回り (クレデンシャルの取得、Target Integrity
の検証、準備完了の通知) と、広告リンク検証という機能の一部を同居させていた。
link-verification モジュールを読んでも機能の全体が見えず、実際に「OP ID と
突き合わせる相手は credentialSubject.id」という同じ規則が content-script.ts と
matching.ts の両方に実装され、片方だけ誤っていた。

link-verification/detect-ad-click.ts へ移し、setupAdClickDetection() として
公開する。setupFrameHandlers への注入ではなく、アプリのエントリで並べて呼ぶ
合成にした。いま拡張機能ごとに変わるものがないため、行使されない seam を作る
理由がない。

detect-ad-click.ts は state.ts に依存しないため、コンテンツスクリプトが
background 専用の状態を引き込むことはない。ビルド成果物で、adClicked が
全フレーム用スクリプトと Service Worker にのみ含まれ、storage.session が
Service Worker にのみ含まれることを確認した。
